+++ qatrack/parts/models.py
@@ -11,8 +11,14 @@ from qatrack.units import models as u_models
 
 class Supplier(models.Model):
 
-    name = models.CharField(max_length=32, unique=True)
+    name = models.CharField(
+        verbose_name=_l("supplier"),
+        max_length=32,
+        unique=True,
+        help_text=_l("Enter a unique name for this supplier"),
+    )
     notes = models.TextField(
+        verbose_name=_l("notes"),
         max_length=255,
         blank=True,
         null=True,
@@ -36,12 +42,17 @@ class Room(models.Model):
 
     site = models.ForeignKey(
         u_models.Site,
+        verbose_name=_l("site"),
         on_delete=models.CASCADE,
         blank=True,
         null=True,
         help_text=_l('Site this storage room is located')
     )
-    name = models.CharField(max_length=32, help_text=_l('Name of room or room number'))
+    name = models.CharField(
+        verbose_name=_l("name"),
+        max_length=32,
+        help_text=_l('Name of room or room number'),
+    )
 
     objects = RoomManager()
 
@@ -73,14 +84,23 @@ class Storage(models.Model):
 
     room = models.ForeignKey(
         Room,
+        verbose_name=_l("room"),
         blank=True,
         null=True,
         help_text=_l('Room for part storage'),
         on_delete=models.CASCADE,
     )
 
-    location = models.CharField(max_length=32, blank=True, null=True)
-    description = models.TextField(max_length=255, null=True, blank=True)
+    location = models.CharField(
+        verbose_name=_l("location"),
+        max_length=32, blank=True, null=True,
+        help_text=_l('Where is this storage located?'),
+    )
+    description = models.TextField(
+        verbose_name=_l("description"),
+        max_length=255, null=True, blank=True,
+        help_text=_l("Optional description of this storage"),
+    )
 
     objects = StorageManager()
 
@@ -104,7 +124,11 @@ class Storage(models.Model):
 
 class PartCategory(models.Model):
 
-    name = models.CharField(max_length=64)
+    name = models.CharField(
+        verbose_name=_l("part category"),
+        max_length=64,
+        help_text=_l("Enter a name for this part category (e.g. Linac Parts, Table Parts etc)"),
+    )
 
     class Meta:
         verbose_name_plural = _l("Categories")
@@ -115,10 +139,22 @@ class PartCategory(models.Model):
 
 class Part(models.Model):
 
-    name = models.CharField(help_text=_l('Brief name describing this part'), max_length=255)
-    part_category = models.ForeignKey(PartCategory, on_delete=models.SET_NULL, blank=True, null=True)
+    name = models.CharField(
+        verbose_name=_l("name"),
+        help_text=_l('Brief name describing this part'),
+        max_length=255,
+    )
+    part_category = models.ForeignKey(
+        PartCategory,
+        verbose_name=_l("part category"),
+        on_delete=models.SET_NULL,
+        blank=True,
+        null=True,
+        help_text=_l("Select what category this part is"),
+    )
     suppliers = models.ManyToManyField(
         Supplier,
+        verbose_name=_l("suppliers"),
         blank=True,
         help_text=_l('Suppliers of this part'),
         related_name='parts',
@@ -126,26 +162,46 @@ class Part(models.Model):
     )
     storage = models.ManyToManyField(
         Storage,
+        verbose_name=_l("storage"),
         through='PartStorageCollection',
         related_name='parts',
         help_text=_l('Storage locations for this part'),
     )
 
-    part_number = models.CharField(max_length=32, unique=True)
+    part_number = models.CharField(
+        verbose_name=_l("part number"),
+        max_length=32,
+        help_text=_l("Enter the part number for this part."),
+    )
+
+    new_or_used = models.CharField(
+        verbose_name=_l("New or Used"),
+        choices=[("both", _l("New & Used")), ("new", _l("New")), ("used", _l("Used"))],
+        default="both",
+        max_length=12,
+        help_text=_l("Select whether this Part is used to track new, used, or both new and used inventory"),
+    )
+
     alt_part_number = models.CharField(
+        verbose_name=_l('Alternate part number'),
         max_length=32,
         blank=True,
         null=True,
-        verbose_name=_l('Alternate part number'),
+        help_text=_l('Is this part also identified by a different number?'),
     )
     quantity_min = models.PositiveIntegerField(
+        verbose_name=_l("Notification level"),
         default=0,
         help_text=_l('Notify when the quantity of this part in storage falls below this number'),
     )
     quantity_current = models.PositiveIntegerField(
-        help_text=_l('The number of parts in storage currently'), default=0, editable=False
+        verbose_name=_l("current quantity"),
+        help_text=_l('The number of parts in storage currently'),
+        default=0,
+        editable=False,
     )
     cost = models.DecimalField(
+        verbose_name=_l("cost"),
         default=0,
         decimal_places=2,
         max_digits=10,
@@ -155,6 +211,7 @@ class Part(models.Model):
         validators=[MinValueValidator(Decimal('0.00'))]
     )
     notes = models.TextField(
+        verbose_name=_l("notes"),
         max_length=255,
         blank=True,
         null=True,
@@ -164,6 +221,9 @@ class Part(models.Model):
 
     class Meta:
         ordering = ['part_number']
+        unique_together = [
+            ('part_number', 'new_or_used'),
+        ]
 
     def __str__(self):
         return '%s%s - %s' % (
@@ -194,10 +254,23 @@ class PartStorageCollectionManager(models.Manager):
 
 class PartStorageCollection(models.Model):
 
-    part = models.ForeignKey(Part, on_delete=models.CASCADE)
-    storage = models.ForeignKey(Storage, on_delete=models.CASCADE)
+    part = models.ForeignKey(
+        Part,
+        verbose_name=_l("part"),
+        on_delete=models.CASCADE,
+        help_text=_l("Select the part to be associated with a Storage"),
+    )
+    storage = models.ForeignKey(
+        Storage,
+        verbose_name=_l("storage"),
+        help_text=_l("Select the Storage to be associated with this Part"),
+        on_delete=models.CASCADE,
+    )
 
-    quantity = models.IntegerField()
+    quantity = models.IntegerField(
+        verbose_name=_l("storage"),
+        help_text=_l("Quantity of parts in this storage"),
+    )
 
     objects = PartStorageCollectionManager()
 
@@ -223,13 +296,24 @@ class PartStorageCollection(models.Model):
 
 class PartSupplierCollection(models.Model):
 
-    part = models.ForeignKey(Part, on_delete=models.CASCADE)
-    supplier = models.ForeignKey(Supplier, on_delete=models.CASCADE)
+    part = models.ForeignKey(
+        Part,
+        verbose_name=_l("part"),
+        help_text=_l("Select the Part to be associated with this Supplier"),
+        on_delete=models.CASCADE,
+    )
+    supplier = models.ForeignKey(
+        Supplier,
+        verbose_name=_l("supplier"),
+        help_text=_l("Select the Supplier to be associated with this Part"),
+        on_delete=models.CASCADE,
+    )
     part_number = models.CharField(
+        verbose_name=_l("part number"),
         max_length=32,
         null=True,
         blank=True,
-        help_text=_l('Does this supplier have a different part number for this part')
+        help_text=_l('Does this supplier have a different part number for this part'),
     )
 
     class Meta:
@@ -239,11 +323,31 @@ class PartSupplierCollection(models.Model):
 
 class PartUsed(models.Model):
 
-    service_event = models.ForeignKey(sl_models.ServiceEvent, on_delete=models.CASCADE)
-    part = models.ForeignKey(Part, help_text=_l('Select the part used'), on_delete=models.CASCADE)
-    from_storage = models.ForeignKey(Storage, null=True, blank=True, on_delete=models.SET_NULL)
+    service_event = models.ForeignKey(
+        sl_models.ServiceEvent,
+        verbose_name=_l("service event"),
+        help_text=_l("Select the Service Event the part was used in"),
+        on_delete=models.CASCADE,
+    )
+    part = models.ForeignKey(
+        Part,
+        verbose_name=_l("part"),
+        help_text=_l('Select the part used'),
+        on_delete=models.CASCADE,
+    )
+    from_storage = models.ForeignKey(
+        Storage,
+        verbose_name=_l("from storage"),
+        help_text=_l('Select which Storage the parts were taken from'),
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+    )
 
-    quantity = models.IntegerField()
+    quantity = models.IntegerField(
+        verbose_name=_l("quantity"),
+        help_text=_l('Select how many parts were used from this Storage'),
+    )
 
     def add_back_to_storage(self):